Add generate-types script
As we’re building on top of the original CAS2’s functionality for CAS2 bail, we
are copying over functionality from the CAS2 repo.

This commit adds the generate-types script and it's dependencies so that we can
use it to keep the types between the API and UI in sync.