Rename esbuild folder to esbuild-configs
As we’re building on top of the original CAS2’s functionality for CAS2 bail, we
are copying over functionality from the CAS2 repo.

In a previous commit, we amended our tsconfig so we can easily reference
'@approved-premises/{types}' that have been generated from the API.

As a result of this change, TS can no longer reference the correct 'esbuild'
module in the local config file. This commit renames the local 'esbuild'
directory to 'esbuild-configs' to avoid any confusion with the node module
'esbuild'.