Add section utils
I have copied over the user utils used to determine which tiles the user should be shown, although because of changes to the local user type, for now we are only returning the POM sections. This will need to be revisited once we understand how we will determine the user’s role.