Refactor code to satisfy suppressImplicitAnyIndexErrors option deprecation
As we’re building on top of the original CAS2’s functionality for CAS2 bail, we
are copying over functionality from the CAS2 repo.

In TypeScript@5.5[1], the `suppressImplicitAnyIndexErrors` compiler option was
deprecated which requires us to refactor our code that relies on accessing
objects using a string index.

In cases where a bigger refactor is needed, I have marked with
`@ts-expect-error Requires refactor to satisfy TS7053` to be looked at a later
date[2].

[1]https://devblogs.microsoft.com/typescript/announcing-typescript-5-5/#disabling-features-deprecated-in-typescript-5.0
[2]https://dsdmoj.atlassian.net/browse/CBA-38?atlOrigin=eyJpIjoiNDkwZWNlNmQxNmVmNGNkMzhkZThiOTI4Nzk3OWZiNTciLCJwIjoiaiJ9