Skip integration tests until testing strategy is established
As we’re building on top of the original CAS2’s functionality for CAS2 bail, we
are copying over functionality from the CAS2 repo.

The CAS2 project uses playwright for e2e tests and cypress for integration
tests. We've decided to add playwright to this project to keep in sync with CAS2.

Therefore, we'll skip integration tests until this work is complete.