Explicitly declare application name in applicationInfo
Pulling this from package.json using relative paths is awkward because when the code is compiled it sits one directory level deeper, so it was causing tests to fail. We could try to do something clever or find a package to give us the project root path, but I don’t think it’s worth it, we may as well just declare the application name explicitly here.