Remove redundant e2e tests
I’m not sure if we will still have the concept of an LCA in our service. We probably will have the concept of an admin, but the tests aren’t working currently and I don’t think we need to worry about maintaining them for MVP. They should be easily to reinstate later provided we have the users with the appropriate permissions in the dev environment.