Delete orphaned CPP details and current offence data
If a user answers yes to the community supervision question and fills in
subsequent screens, the data will persist even after coming back to change the
answer to 'no' and will show on the check your answers screen and be submitted
with the application.

This change ensures we remove that data when the answer is 'no.