Conditional render submission timeline guidance
The timeline for submmissions to be processed differ between court bail
and prison bail, this change ensures the correct guidance is shown.

At the time of writing, the 3 possible application origins are
`courtBail`, `prisonBail` and `homeDetentionCurfew`, and as prison bail
and HDC share the same processing time, we can just check for court
bail.