Temporarily disable retries for e2e tests
Playwright will automatically mark a test as flaky if it passes on
second retry. We currently have a suite of reports e2e tests that are
temporarily broken, which have been flagged as flaky and ultimately
passes the build.

We want these to remain visible until the fix for the tests are in the
API, so as them breaking doesn't stop us from deploying to dev, we'll
temporarily turn off retries for better visibility.