Ensure assess e2e uses recent submitted application
In a previous commit[1], we attempted to mark this test as failing due
to older application data resulting in a 404. However, this solution
didn't work due to the test timing out.

Instead of marking as failing, we ensure the test chooses the most
recent page of submitted applications which lowers the likelihood of
selecting an older application.

A fix should still be implemented in the API to ensure all submitted
applications are viewable.

[1]
https://github.com/ministryofjustice/hmpps-community-accommodation-tier-2-bail-ui/commit/70a5d85492703ba9d6b905a39aaed8b6577fb732#diff-429df44951dd618e2cda8caf7d8af4feca0e6dcdd8bc7b85339ae6c7623dda4dR8-R9