Disable no-param-reassign rule
We have moved over to @ministryofjustice/eslint-config-hmpps, which
currently does not allow for overriding specific rules.

We require the ability to change objects in place as it's a common
pattern in the codebase given the design of our form data is stored.

Therefore, disable this rule on all pages where this occurs.