Remove offence and category and summary from alleged offence
Feedback from testing informed us that users found entering a summary
for each alleged offence tiresome as they usually share the same
context, meaning they would copy and paste the same summary. We'll
instead allow for users to add a summary after adding all the alleged
offences.

This change removes the summary from the alleged offences as well as the
category which will be inferred from the offence title.