Move sqs-hmpps-audit-secret to individual namespaces
As this secret is not set in the test environment, deployments to test
were failing with the error: secret "sqs-hmpps-audit-secret" not found.

We don't require audit logs in the test environment, so moving this
secret to individual namespaces ensures test environment will not try to
retrieve it.

In application code, the call to AWS SQS will silently fail, as
logErrors is set to false by default[1].

[1]
https://github.com/ministryofjustice/hmpps-community-accommodation-tier-2-bail-ui/blob/f943f18755b4c742d2ea8da48f0e04a0c1450fbe/server/services/auditService.ts#L44