Ensure bail hearing date is returned from page response
On some of our pages with dates, the error validation is having the unintended side effect of generating an ISO string on the page body. We don’t have any validation on the bail hearing date page, so that’s not happening here. This is something we will need to revisit, but for now, we return the formatted date from the response method if there are valid date parts in the body, or else we return an empty string.