Add util for to extract bail hearing date from application data
We need this so we can add the bail hearing date to the application as a primary field on submission.

Because at the moment we have no validation on the bail hearing date page, we need to check the date inputs are valid dates to prevent the `dateAndTimeInputsToIsoString` method from erroring. This may not be necessary if/when we have validation on that page.