Allow prison bail referrers to search by CRN
Currently, to start an application for prison bail, the service relies on the following:

1. A Prison Referrer entering the prison number of the applicant,

2. The applicant has a corresponding record in nDelius, and;

3. these two records within Nomis and nDelius are linked (manually or automatically)

Not all applicants will have linked records across nDelius and Nomis,
leading to prison referrers encountering a 404 error when searching via
prison number.

Therefore, we want to allow them search via CRN if they encounter a 404.

Short-term fix:

To mitigate the issue, we are tweaking the error message for all 404s
encountered when searching via prison number to direct the referrer to
the search via CRN page.