Ensure E2E test workflow can be called from API
This ensures the workflow checks out the code from the UI repository
specifically, so it can be run as part of an external workflow, for
instance, for the API pipeline E2E tests.

This also adds a step to send a Slack notification to the relevant slack
channel on failure.