Rename bail information task
Now that we’re removed the support sessions page, it makes sense to rename this task.

This will effectively delete data for the task and cause the task status to reset for all in progress applications, so we will release this early in the morning to minimise disruption to our users.