Add response method to information sources pages
This method allows us to render the information sources data in a user friendly way on the CYA and submitted application views.

NB because the browser submits a string where a single checkboxes is selected, or an array where multiple are selected, I have had to ensure the data is in array format at the start of the response method to avoid unhandled errors.