Add consent refused page
Until now, we’ve handled the path from confirm eligibility -> confirm consent -> task list with a pre task list routing util. But now that we want more flexibility for users to answer ‘no’ on the consent page and return to the task list to continue the application, it makes sense to recreate the ‘consent refused’ page within the task, rather than as its own page independent of the task list system.