Update confirm consent page
We want the task to display as ‘In progress’ if the user answers ‘no’, to make sure they come back later to confirm that consent has been given. This is not straightforward to achieve, so I’ve gone for the slightly hacky solution of saving the page body as an empty object if hasGivenConsent is ‘no’. This effectively means that the user cannot persist a ‘no’ answer on an application. But it does mean that if they continue with the application, this task will be marked as ‘In progress’ for them to return to. This should achieve what we want to achieve in almost all cases - for now we aren’t expecting many referrers to complete an application if the answer here is truly ‘no’.