Remove consent refused reason question
Given that we are preventing the user from persisting a ‘no’ answer, it would be strange to persist the reason. This data may be useful at some point but until now we haven’t used it, so we’re removing it to achieve the desired flow for the consent task.