Add links to interview question sheets
In case parts of the digital service are down, we want referrers to be able to access a version of the question set to collect information offline, so we’re providing links from the homepage to a HTML page and docx document for that purpose.

As these documents are stored within our repo, if they are updated in future they must also be updated here.