Check if ID documents field is defined before chaining includes method
When we made the changes to the confirm consent page, we moved the page.next() call in the pagesController update method above the applicationService.save() method: https://github.com/ministryofjustice/hmpps-community-accommodation-tier-2-bail-ui/blob/bf8a917c9f2ffe61cac1d528b08b692f42ecc24c/server/controllers/apply/applications/pagesController.ts#L55.

We did this so that on the consent page we could work out the next page before removing the page data if the referrer answers no. However, in the case of the applicant ID page, page.next() relies on page.body.idDocuments being defined, which it will not be if no answer has been selected. And yet page.next() will be called before page.errors(), which is called inside applicationService.save().

This check will solve that problem, and is the only place I can see the issue in the codebase. However, it’s a fix that may need to be repeated in future cases. Therefore we may well revisit this shortly with a different solution.