Ensure page.next() is called before page save only for confirm consent page
For the confirm consent page, we want to work out which the next page should be before we save, because for this page only, if the answer is ‘no’, we remove the answer on save (so that we can retain the task status as ‘in progress’). However, we can’t do this for every page because other pages may rely on defined body properties when calling page.next(), which will cause the service to error.

So to avoid this risk, we will change the function call order only in the case of the confirm consent page.