Refactor other relevance check pages
We want to ensure that the cover and details pages will be presented together on check your answers. To do this we have an elaborate response method on the cover page which will pull application data related to the details page. In previous cases we were manually skipping the details pages in the CYA utils, but I think the approach in this PR of returning an empty object from the response method is more future proof and less brittle. So I have extended it to the brain injury and communication and language cases here.