Remove OASys information source guard
We added this guard when we moved from a single OASys information source to two separate ones, in case any in progress applications had selected that option. The guard ensures that that OASys option would render in CYA and the submitted applications views.

It has been two weeks since we implemented that change and it is now unlikely that any genuinely in progress applications are using that option, so we can remove the guard.