Update contract test framework
This enables the contract testing to handle the new API spec, which is
of version 3.1.0, which swagger-mock-validator was not compatible with.

The contract tests require access to the API to download the openAPI
spec, and therefore need to be run on a self-hosted runner. This isn't
possible using the predefined workflow from the hmpps-github-actions
repository, so it has been copied over and updated to retain existing
functionality.